--- api/app/src/test/kotlin/packit/TestConfig.kt
@@ -1,6 +1,0 @@
-package packit
-
-import org.springframework.boot.test.context.TestConfiguration
-
-@TestConfiguration
-class TestConfig
