+++ api/app/src/test/kotlin/packit/integration/IntegrationTest.kt
@@ -2,12 +2,14 @@ package packit.integration
 
 import org.assertj.core.api.Assertions.assertThat
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.boot.test.web.client.TestRestTemplate
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
 import kotlin.test.assertEquals
 
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 abstract class IntegrationTest
 {
     @Autowired
@@ -20,3 +22,4 @@ abstract class IntegrationTest
         assertThat(responseEntity.body).isNotEmpty
     }
 }
+
