+++ api/app/src/test/kotlin/packit/integration/controllers/PacketControllerTest.kt
@@ -1,10 +1,8 @@
 package packit.integration.controllers
 
 import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
 import packit.integration.IntegrationTest
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class PacketControllerTest : IntegrationTest()
 {
     @Test
